+++ cookbook/case_studies/ml_training/spark_horovod/Makefile
@@ -1,0 +1,3 @@
+PREFIX=spark_horovod
+include ../../../common/common.mk
+include ../../../common/leaf.mk
