+++ cookbook/case_studies/ml_training/spark_horovod/README.rst
@@ -1,0 +1,8 @@
+.. _spark_horovod:
+
+Horovod
+=========
+
+Horovod is a distributed deep learning training framework for TensorFlow, Keras, PyTorch, and Apache MXNet. 
+The goal of Horovod is to make distributed deep learning fast and easy to use. It uses the all-reduce algorithm for fast 
+distributed training rather than a centralized parameter server approach (all-reduce vs. parameter server).
