+++ cookbook/case_studies/ml_training/spark_horovod/requirements.in
@@ -1,0 +1,1 @@
+flytekitplugins-spark>=0.16.0
