+++ cookbook/case_studies/ml_training/spark_horovod/sandbox.config
@@ -1,0 +1,3 @@
+[sdk]
+workflow_packages=spark_horovod
+python_venv=flytekit_venv
