--- cookbook/integrations/external_services/spark_horovod/Makefile
@@ -1,3 +1,0 @@
-PREFIX=spark_horovod
-include ../../../common/common.mk
-include ../../../common/leaf.mk
