--- cookbook/integrations/external_services/spark_horovod/requirements.in
@@ -1,2 +1,0 @@
-git+https://github.com/flyteorg/flytekit.git@ac3c84b170cfc658ad78e87e1b2d0da0cfe2f6f7#subdirectory=plugins/flytekit-kf-mpi
-flytekitplugins-spark>=0.16.0
