--- cookbook/integrations/external_services/spark_horovod/sandbox.config
@@ -1,3 +1,0 @@
-[sdk]
-workflow_packages=horovod
-python_venv=flytekit_venv
